% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{categorias_bp}
\alias{categorias_bp}
\title{Lista de categorias contabeis utilizadas internamente}
\format{
Uma lista nomeada com vetores de caracteres.
}
\usage{
categorias_bp
}
\description{
Este objeto lista, em formato de `list`, as categorias contabeis utilizadas
para classificacao de contas em funcoes internas do pacote.
}
\details{
As chaves representam os grupos (por exemplo, ACF = Ativo Circulante Financeiro),
e os vetores associados contem os nomes das contas que pertencem a cada grupo.
}
\keyword{internal}
