% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{classificar_conta_bp}
\alias{classificar_conta_bp}
\title{Classifica contas contabeis de acordo com categorias predefinidas}
\usage{
classificar_conta_bp(conta)
}
\arguments{
\item{conta}{Um vetor de caracteres com o(s) nome(s) da(s) conta(s) a classificar.}
}
\value{
Um vetor de caracteres com a categoria correspondente ou \code{NA} se nao classificada.
}
\description{
A funcao recebe o nome de uma conta e retorna sua categoria contabil conforme a lista `categorias`.
}
\keyword{internal}
