% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{conv_br_numeric}
\alias{conv_br_numeric}
\title{Converte texto contabil brasileiro para numero negativo quando necessario}
\usage{
conv_br_numeric(vetor)
}
\arguments{
\item{vetor}{Vetor de caracteres contendo os valores a serem convertidos.}
}
\value{
Vetor numerico com os valores convertidos.
}
\description{
Esta funcao transforma strings representando valores contabeis no formato brasileiro
(com virgula decimal, ponto como separador de milhar e uso de parenteses para indicar valor negativo)
em valores numericos padrao do R.
}
\keyword{internal}
