% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{normalizar_texto}
\alias{normalizar_texto}
\title{Normaliza texto removendo acentos e cedilha}
\usage{
normalizar_texto(x)
}
\arguments{
\item{x}{Vetor de caracteres a ser normalizado.}
}
\value{
Vetor de caracteres normalizado.
}
\description{
Esta funcao converte o texto para minusculas, normaliza para Unicode NFC
e remove todos os acentos, cedilhas e marcas diacriticas, retornando
apenas caracteres ASCII basicos.
}
\examples{
normalizar_texto(c("Ção", "Ótimo", "maçã", "PÃO"))
# [1] "cao"   "otimo" "maca" "pao"
}
