% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{normalizar_texto}
\alias{normalizar_texto}
\title{Normaliza texto removendo acentos e cedilha}
\usage{
normalizar_texto(x)
}
\arguments{
\item{x}{Vetor de caracteres a ser normalizado.}
}
\value{
Vetor de caracteres normalizado.
}
\description{
Função interna utilizada nos processos de padronização textual.
Não deve ser chamada diretamente pelo usuário final.
}
\keyword{internal}
